Set Apache user and group to vagrant
In order for Apache to be able to overwrite files that it creates, it
needs to be running as the same user as the owner of the directory in
question. In our case, the directory is within /vagrant and Vagrant sets
the ownership to vagrant:vagrant. Hence, we need to make Apache run as
the vagrant user.